+++ os/StarryOS/kernel/src/pseudofs/proc.rs
@@ -909,9 +909,9 @@ impl SimpleDirOps for ThreadFdDir {
             Ok(None) => return Box::new(iter::empty()),
             Err(error) => panic!("procfs fd directory has an invalid user extension: {error}"),
         };
-        let scope = task.as_thread().scope.read();
-        let fd_table = FD_TABLE.scope_cell(&scope).clone();
-        drop(scope);
+        let fd_table = task
+            .as_thread()
+            .with_scope(|scope| FD_TABLE.scope_cell(scope).clone());
         let ids = fd_table
             .read()
             .ids()
@@ -924,9 +924,9 @@ impl SimpleDirOps for ThreadFdDir {
         let fs = self.fs.clone();
         let task = require_proc_task(&self.task)?;
         let fd = name.parse::<u32>().map_err(|_| VfsError::NotFound)?;
-        let scope = task.as_thread().scope.read();
-        let fd_table = FD_TABLE.scope_cell(&scope).clone();
-        drop(scope);
+        let fd_table = task
+            .as_thread()
+            .with_scope(|scope| FD_TABLE.scope_cell(scope).clone());
         let path = fd_table
             .read()
             .get(fd as _)
@@ -1353,9 +1353,9 @@ impl SimpleDirOps for ThreadDir {
                 let task = self.task;
                 SimpleFile::new_regular(fs, move || {
                     let task = require_proc_task(&task)?;
-                    let scope = task.as_thread().scope.read();
-                    let ctx_arc = FS_CONTEXT.scope_cell(&scope).clone();
-                    drop(scope);
+                    let ctx_arc = task
+                        .as_thread()
+                        .with_scope(|scope| FS_CONTEXT.scope_cell(scope).clone());
                     let ctx = ctx_arc.lock();
                     Ok(crate::pseudofs::proc_mountinfo::render_mounts(&ctx))
                 })
@@ -1365,9 +1365,9 @@ impl SimpleDirOps for ThreadDir {
                 let task = self.task;
                 SimpleFile::new_regular(fs, move || {
                     let task = require_proc_task(&task)?;
-                    let scope = task.as_thread().scope.read();
-                    let ctx_arc = FS_CONTEXT.scope_cell(&scope).clone();
-                    drop(scope);
+                    let ctx_arc = task
+                        .as_thread()
+                        .with_scope(|scope| FS_CONTEXT.scope_cell(scope).clone());
                     let ctx = ctx_arc.lock();
                     Ok(crate::pseudofs::proc_mountinfo::render_mountinfo(&ctx))
                 })