+++ os/StarryOS/kernel/src/syscall/task/namespace.rs
@@ -257,9 +257,9 @@ fn setns_via_pidfd(pidfd: &PidFd, nstype: u32) -> AxResult<isize> {
     let target_proc = pidfd.process_data()?;
     let target_mnt_fs_ns = if nstype & CLONE_NEWNS != 0 {
         let task = get_task(target_proc.proc.pid())?;
-        let scope = task.as_thread().scope.read();
-        let fs_context = FS_CONTEXT.scope_cell(&scope).clone();
-        drop(scope);
+        let fs_context = task
+            .as_thread()
+            .with_scope(|scope| FS_CONTEXT.scope_cell(scope).clone());
         Some(fs_context.lock().mount_namespace().clone())
     } else {
         None